requirements.txt
connexion
gunicorn
swagger-ui
swagger-ui-bundle
psycopg2-binary
requests
Flask
oauthlib
pyOpenSSL
Flask-Login
flask_restful
webargs
numpy
pandas
gspread
sklearn
wordcloud
matplotlib
authlib
python-dotenv